refactor based on disucssion
